requirements.txt
ansi2html
ansible
ansible-compat
ansible-core
ansible-lint
api-client
api-client-pydantic
#apiclient-pydantic-generator
#appnope
argcomplete
#argon2-cffi
#argon2-cffi-bindings
#arrow
#asttokens
#attrs
#awxkit
#backcall
beautifulsoup4
#binaryornot
#black
#bleach
#bracex
#cerberus
certifi
cffi
#chardet
#charset-normalizer
click
click-help-colors
colorama
commonmark
configparser
cookiecutter
cryptography
datamodel-code-generator
debugpy
decorator
defusedxml
distlib
dnspython
docker
docopt
dump
#email-validator
enrich
entrypoints
executing
fastjsonschema
filelock
flake8
genson
idna
inflect
iniconfig
ipykernel
ipython
ipython-genutils
ipywidgets
isodate
isort
jedi
Jinja2
jinja2-time
jmespath
jsonschema
jupyter
#jupyter-client
#jupyter-console
#jupyter-core
#jupyterlab-pygments
#jupyterlab-widgets
loguru
MarkupSafe
matplotlib-inline
#mccabe
mistune
molecule
molecule-docker
molecule-plugins
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
netaddr
notebook
ntlm-auth
#openapi-schema-validator
#openapi-spec-validator
packaging
pandocfilters
parso
passlib
pathspec
pdf2image
pexpect
pickleshare
pillow
pipdeptree
platformdirs
pluggy
prance
#prometheus-client
#prompt-toolkit
psutil
ptyprocess
#pure-eval
py
pyasn1
#pyasn1-modules
pycodestyle
pycparser
pydantic
pyflakes
Pygments
pyparsing
#pyrsistent
PySnooper
pyspnego
pytest
pytest-testinfra
python-dateutil
python-ldap
python-slugify
python-vagrant
pyvim
pyvmomi
pywinrm
PyYAML
pyzmq
#qtconsole
#QtPy
questionary
requests
requests-credssp
#requests-ntlm
resolvelib
rich
ruamel.yaml
ruamel.yaml.clib
Send2Trash
shellingham
six
#soupsieve
stack-data
#steampunk-spotter
stringcase
subprocess-tee
tenacity
terminado
text-unidecode
tinycss2
toml
tomli
tornado
traitlets
#typed-ast
#typer
#typing_extensions
urllib3
virtualenv
#wcmatch
#wcwidth
webencodings
websocket-client
widgetsnbextension
xmltodict
yamllint
#yapf
